The dependencies test works!!
